requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
django-mptt==0.13.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Pillow==9.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
